Move NodeLink class to separate file and refactor slightly.
